+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   * `CamelTo (separator :: Symbol)` and its variants, `CamelToSnake` and `CamelToKebab`: Transform field names written
   in CamelCase to snake\_case, kebab-case, or with a custom separator.
 * Remove redundant metadata about the query nature when logging the query
+* Add `selectOneWhereIn` that can match a row's column in a value of user-provided values
+
 ## 0.0.1.0 -- 2021-11-05
 
 * First version. Released on an unsuspecting world.